pyproject.toml
[project]
name = "photo-keyword-tagger"
version = "0.1.0"
description = "A Python package for managing Lightroom catalogues by adding keywords to collections of photos"
readme = "README.md"
authors = [
    { name = "Gurudev Ilangovan", email = "ilangurudev@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "tqdm>=4.66.0",
    "click>=8.0.0"
]

[project.scripts]
photo-keyword-tagger = "photo_keyword_tagger.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "ipykernel>=7.1.0"
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["photo_keyword_tagger"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
